plugins/janus_reverse_plugin/requirements.txt
git+https://github.com/deepseek-ai/Janus.git
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.25.0
tqdm>=4.60.0
